Move type data from spi to types-transfer
