+++ portability-types-transfer/src/main/java/org/dataportabilityproject/types/transfer/models/mail/MailMessageModel.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.spi.transfer.models.mail;
+package org.dataportabilityproject.types.transfer.models.mail;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
