+++ portability-types-transfer/src/main/java/org/dataportabilityproject/types/transfer/models/tasks/TaskListModel.java
@@ -13,34 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.spi.transfer.models.photos;
-
+package org.dataportabilityproject.types.transfer.models.tasks;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import org.dataportabilityproject.types.transfer.models.DataModel;
 
-public class PhotoAlbum {
+public class TaskListModel extends DataModel {
     private final String id;
     private final String name;
-    private final String description;
 
-    /** The {@code id} is used to associate photos with this album. **/
     @JsonCreator
-    public PhotoAlbum(
+    public TaskListModel(
         @JsonProperty("id") String id,
-        @JsonProperty("name") String name,
-        @JsonProperty("description") String description) {
+        @JsonProperty("name") String name) {
         this.id = id;
         this.name = name;
-        this.description = description;
     }
 
     public String getName() {
         return name;
-    }
-
-    public String getDescription() {
-        return description;
     }
 
     public String getId() {
