+++ portability-types-transfer/src/main/java/org/dataportabilityproject/types/transfer/models/tasks/TaskModel.java
@@ -13,36 +13,35 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.spi.transfer.models.calendar;
-
+package org.dataportabilityproject.types.transfer.models.tasks;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class CalendarAttendeeModel {
-    private final String displayName;
-    private final String email;
-    private final boolean optional;
+public class TaskModel {
+    private final String taskListId;
+    private final String text;
+    private final String notes;
 
     @JsonCreator
-    public CalendarAttendeeModel(
-        @JsonProperty("displayName") String displayName,
-        @JsonProperty("email") String email,
-        @JsonProperty("optional") boolean optional) {
-        this.displayName = displayName;
-        this.email = email;
-        this.optional = optional;
+    public TaskModel(
+        @JsonProperty("taskListId") String taskListId,
+        @JsonProperty("text") String text,
+        @JsonProperty("notes") String notes) {
+        this.taskListId = taskListId;
+        this.text = text;
+        this.notes = notes;
     }
 
-    public boolean getOptional() {
-        return optional;
+    public String getText() {
+        return text;
     }
 
-    public String getEmail() {
-        return email;
+    public String getNotes() {
+        return notes;
     }
 
-    public String getDisplayName() {
-        return displayName;
+    public String getTaskListId() {
+        return taskListId;
     }
 }
