+++ portability-types-transfer/src/test/java/org/dataportabilityproject/types/transfer/models/photos/PhotosModelWrapperTest.java
@@ -1,12 +1,9 @@
-package org.dataportabilityproject.spi.transfer.models.photos;
+package org.dataportabilityproject.types.transfer.models.photos;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableList;
 import com.google.common.truth.Truth;
 import java.util.List;
-import org.dataportabilityproject.spi.transfer.models.mail.MailContainerModel;
-import org.dataportabilityproject.spi.transfer.models.mail.MailMessageModel;
-import org.dataportabilityproject.spi.transfer.models.mail.MailModelWrapper;
 import org.dataportabilityproject.types.transfer.models.DataModel;
 import org.junit.Test;
